# CUDA: revise q8_1 data layout for mul_mat_q