This PR changes the data layout for the FP32 values quantized to q8_1 in conjunction with MMQ. The blocks with 32 values are consolidated into larger blocks of 128 values in order to make them align to 16 bits and have the same data layout in global and shared memory. This is relevant for asynchronous data loading (Ampere or newer, not in this PR). The memory layout in shared memory now also allows for more efficient data loading for int8 tensor cores; for optimal performance you need an offset of 16 bytes between columns and it just so happens that for 128 values this is the exact size that the q8_1 scales take up.

The q8_1 memory layout in global memory for MMQ is that values are first consolidated into blocks of size 128 along ne10. These blocks are then treated as individual elements and transposed.

In `ggml-cuda.cu` I changed the argument for whether or not src1 should be quantized to a function pointer to the specific function that should be used. The data transfer for `--split-mode row` needs to be different. I added a utility function `cudaMemcpy2DPeerAsync` since it for whatever reason does not exist in the CUDA toolkit.

Performance on RTX4090/RTX3090 stays mostly the same, the changes for P40/RX 6800 are larger. The performance increases on average. 

<details>
<summary>Specific numbers</summary>

| GPU        | Model             |     Microbatch size | FlashAttention     | Test     |     t/s master |     t/s cuda-mmq-q8_1-2 |     Speedup |
| :--------- | :---------------- | ------------------: | :----------------- | :------- | -------------: | ----------------------: | ----------: |
| RTX 4090   | llama 8B Q2_K_M   |                  16 | Yes                | pp512    |        1102.82 |                 1118.51 |        1.01 |
| RTX 4090   | llama 8B Q2_K_M   |                  32 | Yes                | pp512    |        1562.51 |                 1584.70 |        1.01 |
| RTX 4090   | llama 8B Q2_K_M   |                  64 | Yes                | pp512    |        2022.17 |                 2033.73 |        1.01 |
| RTX 4090   | llama 8B Q2_K_M   |                 128 | Yes                | pp512    |        3726.70 |                 3728.89 |        1.00 |
| RTX 4090   | llama 8B Q2_K_M   |                 256 | Yes                | pp512    |        5989.97 |                 5990.15 |        1.00 |
| RTX 4090   | llama 8B Q2_K_M   |                 512 | Yes                | pp512    |        7825.63 |                 7816.57 |        1.00 |
| RTX 4090   | llama 8B Q3_K_S   |                  16 | Yes                | pp512    |         888.89 |                  887.58 |        1.00 |
| RTX 4090   | llama 8B Q3_K_S   |                  32 | Yes                | pp512    |        1342.94 |                 1367.00 |        1.02 |
| RTX 4090   | llama 8B Q3_K_S   |                  64 | Yes                | pp512    |        1878.46 |                 1895.10 |        1.01 |
| RTX 4090   | llama 8B Q3_K_S   |                 128 | Yes                | pp512    |        3627.84 |                 3626.18 |        1.00 |
| RTX 4090   | llama 8B Q3_K_S   |                 256 | Yes                | pp512    |        5881.27 |                 5882.97 |        1.00 |
| RTX 4090   | llama 8B Q3_K_S   |                 512 | Yes                | pp512    |        7728.16 |                 7744.08 |        1.00 |
| RTX 4090   | llama 8B Q4_0     |                  16 | Yes                | pp512    |        1598.48 |                 1642.08 |        1.03 |
| RTX 4090   | llama 8B Q4_0     |                  32 | Yes                | pp512    |        2616.94 |                 2644.69 |        1.01 |
| RTX 4090   | llama 8B Q4_0     |                  64 | Yes                | pp512    |        3693.61 |                 3764.31 |        1.02 |
| RTX 4090   | llama 8B Q4_0     |                 128 | Yes                | pp512    |        3580.46 |                 3577.67 |        1.00 |
| RTX 4090   | llama 8B Q4_0     |                 256 | Yes                | pp512    |        5838.55 |                 5839.45 |        1.00 |
| RTX 4090   | llama 8B Q4_0     |                 512 | Yes                | pp512    |        7775.65 |                 7807.98 |        1.00 |
| RTX 4090   | llama 8B Q4_1     |                  16 | Yes                | pp512    |        1560.23 |                 1568.59 |        1.01 |
| RTX 4090   | llama 8B Q4_1     |                  32 | Yes                | pp512    |        2574.10 |                 2563.09 |        1.00 |
| RTX 4090   | llama 8B Q4_1     |                  64 | Yes                | pp512    |        3670.49 |                 3702.49 |        1.01 |
| RTX 4090   | llama 8B Q4_1     |                 128 | Yes                | pp512    |        3535.60 |                 3538.59 |        1.00 |
| RTX 4090   | llama 8B Q4_1     |                 256 | Yes                | pp512    |        5785.55 |                 5788.89 |        1.00 |
| RTX 4090   | llama 8B Q4_1     |                 512 | Yes                | pp512    |        7658.33 |                 7668.97 |        1.00 |
| RTX 4090   | llama 8B Q4_K_S   |                  16 | Yes                | pp512    |        1392.65 |                 1409.09 |        1.01 |
| RTX 4090   | llama 8B Q4_K_S   |                  32 | Yes                | pp512    |        2367.72 |                 2405.85 |        1.02 |
| RTX 4090   | llama 8B Q4_K_S   |                  64 | Yes                | pp512    |        3131.95 |                 3163.14 |        1.01 |
| RTX 4090   | llama 8B Q4_K_S   |                 128 | Yes                | pp512    |        3582.14 |                 3583.80 |        1.00 |
| RTX 4090   | llama 8B Q4_K_S   |                 256 | Yes                | pp512    |        5850.85 |                 5865.58 |        1.00 |
| RTX 4090   | llama 8B Q4_K_S   |                 512 | Yes                | pp512    |        7783.61 |                 7789.94 |        1.00 |
| RTX 4090   | llama 8B Q5_0     |                  16 | Yes                | pp512    |        1257.68 |                 1252.06 |        1.00 |
| RTX 4090   | llama 8B Q5_0     |                  32 | Yes                | pp512    |        1988.88 |                 1984.19 |        1.00 |
| RTX 4090   | llama 8B Q5_0     |                  64 | Yes                | pp512    |        2919.13 |                 2956.97 |        1.01 |
| RTX 4090   | llama 8B Q5_0     |                 128 | Yes                | pp512    |        3480.97 |                 3479.59 |        1.00 |
| RTX 4090   | llama 8B Q5_0     |                 256 | Yes                | pp512    |        5729.44 |                 5731.15 |        1.00 |
| RTX 4090   | llama 8B Q5_0     |                 512 | Yes                | pp512    |        7690.57 |                 7696.39 |        1.00 |
| RTX 4090   | llama 8B Q5_1     |                  16 | Yes                | pp512    |        1303.03 |                 1315.82 |        1.01 |
| RTX 4090   | llama 8B Q5_1     |                  32 | Yes                | pp512    |        2156.25 |                 2044.06 |        0.95 |
| RTX 4090   | llama 8B Q5_1     |                  64 | Yes                | pp512    |        2991.30 |                 2888.40 |        0.97 |
| RTX 4090   | llama 8B Q5_1     |                 128 | Yes                | pp512    |        3494.70 |                 3490.87 |        1.00 |
| RTX 4090   | llama 8B Q5_1     |                 256 | Yes                | pp512    |        5709.80 |                 5707.48 |        1.00 |
| RTX 4090   | llama 8B Q5_1     |                 512 | Yes                | pp512    |        7644.22 |                 7655.42 |        1.00 |
| RTX 4090   | llama 8B Q5_K_S   |                  16 | Yes                | pp512    |        1194.83 |                 1271.99 |        1.06 |
| RTX 4090   | llama 8B Q5_K_S   |                  32 | Yes                | pp512    |        1913.44 |                 2047.03 |        1.07 |
| RTX 4090   | llama 8B Q5_K_S   |                  64 | Yes                | pp512    |        2695.83 |                 2800.62 |        1.04 |
| RTX 4090   | llama 8B Q5_K_S   |                 128 | Yes                | pp512    |        3547.30 |                 3544.97 |        1.00 |
| RTX 4090   | llama 8B Q5_K_S   |                 256 | Yes                | pp512    |        5789.94 |                 5790.79 |        1.00 |
| RTX 4090   | llama 8B Q5_K_S   |                 512 | Yes                | pp512    |        7723.31 |                 7725.42 |        1.00 |
| RTX 4090   | llama 8B Q6_K     |                  16 | Yes                | pp512    |        1096.41 |                 1125.48 |        1.03 |
| RTX 4090   | llama 8B Q6_K     |                  32 | Yes                | pp512    |        1839.73 |                 1783.85 |        0.97 |
| RTX 4090   | llama 8B Q6_K     |                  64 | Yes                | pp512    |        2574.44 |                 2584.98 |        1.00 |
| RTX 4090   | llama 8B Q6_K     |                 128 | Yes                | pp512    |        3456.14 |                 3457.61 |        1.00 |
| RTX 4090   | llama 8B Q6_K     |                 256 | Yes                | pp512    |        5640.23 |                 5644.63 |        1.00 |
| RTX 4090   | llama 8B Q6_K     |                 512 | Yes                | pp512    |        7442.21 |                 7445.44 |        1.00 |
| RTX 4090   | llama 8B Q8_0     |                  16 | Yes                | pp512    |        1036.27 |                 1054.82 |        1.02 |
| RTX 4090   | llama 8B Q8_0     |                  32 | Yes                | pp512    |        1795.01 |                 1973.04 |        1.10 |
| RTX 4090   | llama 8B Q8_0     |                  64 | Yes                | pp512    |        2988.63 |                 3063.31 |        1.02 |
| RTX 4090   | llama 8B Q8_0     |                 128 | Yes                | pp512    |        3377.35 |                 3373.79 |        1.00 |
| RTX 4090   | llama 8B Q8_0     |                 256 | Yes                | pp512    |        5574.59 |                 5571.14 |        1.00 |
| RTX 4090   | llama 8B Q8_0     |                 512 | Yes                | pp512    |        7515.20 |                 7524.32 |        1.00 |
| RTX 3090   | llama 8B Q2_K_M   |                  16 | Yes                | pp512    |         530.92 |                  528.76 |        1.00 |
| RTX 3090   | llama 8B Q2_K_M   |                  32 | Yes                | pp512    |         630.30 |                  621.68 |        0.99 |
| RTX 3090   | llama 8B Q2_K_M   |                  64 | Yes                | pp512    |         918.81 |                  895.11 |        0.97 |
| RTX 3090   | llama 8B Q2_K_M   |                 128 | Yes                | pp512    |        1159.97 |                 1138.32 |        0.98 |
| RTX 3090   | llama 8B Q2_K_M   |                 256 | Yes                | pp512    |        1502.26 |                 1464.59 |        0.97 |
| RTX 3090   | llama 8B Q2_K_M   |                 512 | Yes                | pp512    |        1541.40 |                 1562.12 |        1.01 |
| RTX 3090   | llama 8B Q3_K_S   |                  16 | Yes                | pp512    |         419.23 |                  415.71 |        0.99 |
| RTX 3090   | llama 8B Q3_K_S   |                  32 | Yes                | pp512    |         539.39 |                  527.62 |        0.98 |
| RTX 3090   | llama 8B Q3_K_S   |                  64 | Yes                | pp512    |         858.80 |                  841.15 |        0.98 |
| RTX 3090   | llama 8B Q3_K_S   |                 128 | Yes                | pp512    |        1203.35 |                 1170.51 |        0.97 |
| RTX 3090   | llama 8B Q3_K_S   |                 256 | Yes                | pp512    |        1555.93 |                 1560.10 |        1.00 |
| RTX 3090   | llama 8B Q3_K_S   |                 512 | Yes                | pp512    |        1652.37 |                 1663.21 |        1.01 |
| RTX 3090   | llama 8B Q4_0     |                  16 | Yes                | pp512    |         949.48 |                  939.88 |        0.99 |
| RTX 3090   | llama 8B Q4_0     |                  32 | Yes                | pp512    |        1267.78 |                 1264.83 |        1.00 |
| RTX 3090   | llama 8B Q4_0     |                  64 | Yes                | pp512    |        1737.34 |                 1726.78 |        0.99 |
| RTX 3090   | llama 8B Q4_0     |                 128 | Yes                | pp512    |        2054.06 |                 2052.54 |        1.00 |
| RTX 3090   | llama 8B Q4_0     |                 256 | Yes                | pp512    |        2491.75 |                 2500.63 |        1.00 |
| RTX 3090   | llama 8B Q4_0     |                 512 | Yes                | pp512    |        2620.33 |                 2604.94 |        0.99 |
| RTX 3090   | llama 8B Q4_1     |                  16 | Yes                | pp512    |         925.82 |                  920.58 |        0.99 |
| RTX 3090   | llama 8B Q4_1     |                  32 | Yes                | pp512    |        1239.31 |                 1217.35 |        0.98 |
| RTX 3090   | llama 8B Q4_1     |                  64 | Yes                | pp512    |        1623.27 |                 1625.47 |        1.00 |
| RTX 3090   | llama 8B Q4_1     |                 128 | Yes                | pp512    |        1955.09 |                 1963.57 |        1.00 |
| RTX 3090   | llama 8B Q4_1     |                 256 | Yes                | pp512    |        2352.57 |                 2298.17 |        0.98 |
| RTX 3090   | llama 8B Q4_1     |                 512 | Yes                | pp512    |        2479.74 |                 2461.23 |        0.99 |
| RTX 3090   | llama 8B Q4_K_S   |                  16 | Yes                | pp512    |         738.96 |                  741.87 |        1.00 |
| RTX 3090   | llama 8B Q4_K_S   |                  32 | Yes                | pp512    |         980.60 |                  984.92 |        1.00 |
| RTX 3090   | llama 8B Q4_K_S   |                  64 | Yes                | pp512    |        1466.61 |                 1436.16 |        0.98 |
| RTX 3090   | llama 8B Q4_K_S   |                 128 | Yes                | pp512    |        1742.16 |                 1737.24 |        1.00 |
| RTX 3090   | llama 8B Q4_K_S   |                 256 | Yes                | pp512    |        2123.83 |                 2140.76 |        1.01 |
| RTX 3090   | llama 8B Q4_K_S   |                 512 | Yes                | pp512    |        2277.97 |                 2307.83 |        1.01 |
| RTX 3090   | llama 8B Q5_0     |                  16 | Yes                | pp512    |         657.66 |                  648.62 |        0.99 |
| RTX 3090   | llama 8B Q5_0     |                  32 | Yes                | pp512    |         935.54 |                  916.87 |        0.98 |
| RTX 3090   | llama 8B Q5_0     |                  64 | Yes                | pp512    |        1359.57 |                 1305.98 |        0.96 |
| RTX 3090   | llama 8B Q5_0     |                 128 | Yes                | pp512    |        1782.90 |                 1758.54 |        0.99 |
| RTX 3090   | llama 8B Q5_0     |                 256 | Yes                | pp512    |        2245.18 |                 2206.38 |        0.98 |
| RTX 3090   | llama 8B Q5_0     |                 512 | Yes                | pp512    |        2380.48 |                 2394.16 |        1.01 |
| RTX 3090   | llama 8B Q5_1     |                  16 | Yes                | pp512    |         723.15 |                  712.35 |        0.99 |
| RTX 3090   | llama 8B Q5_1     |                  32 | Yes                | pp512    |         938.94 |                  949.47 |        1.01 |
| RTX 3090   | llama 8B Q5_1     |                  64 | Yes                | pp512    |        1368.83 |                 1356.26 |        0.99 |
| RTX 3090   | llama 8B Q5_1     |                 128 | Yes                | pp512    |        1753.49 |                 1657.81 |        0.95 |
| RTX 3090   | llama 8B Q5_1     |                 256 | Yes                | pp512    |        2186.09 |                 2136.65 |        0.98 |
| RTX 3090   | llama 8B Q5_1     |                 512 | Yes                | pp512    |        2312.84 |                 2291.97 |        0.99 |
| RTX 3090   | llama 8B Q5_K_S   |                  16 | Yes                | pp512    |         598.35 |                  643.18 |        1.07 |
| RTX 3090   | llama 8B Q5_K_S   |                  32 | Yes                | pp512    |         856.19 |                  851.94 |        1.00 |
| RTX 3090   | llama 8B Q5_K_S   |                  64 | Yes                | pp512    |        1308.06 |                 1234.13 |        0.94 |
| RTX 3090   | llama 8B Q5_K_S   |                 128 | Yes                | pp512    |        1591.37 |                 1593.33 |        1.00 |
| RTX 3090   | llama 8B Q5_K_S   |                 256 | Yes                | pp512    |        1994.02 |                 1976.63 |        0.99 |
| RTX 3090   | llama 8B Q5_K_S   |                 512 | Yes                | pp512    |        2135.53 |                 2143.79 |        1.00 |
| RTX 3090   | llama 8B Q6_K     |                  16 | Yes                | pp512    |         564.58 |                  565.88 |        1.00 |
| RTX 3090   | llama 8B Q6_K     |                  32 | Yes                | pp512    |         825.75 |                  822.66 |        1.00 |
| RTX 3090   | llama 8B Q6_K     |                  64 | Yes                | pp512    |        1230.89 |                 1239.15 |        1.01 |
| RTX 3090   | llama 8B Q6_K     |                 128 | Yes                | pp512    |        1541.75 |                 1536.46 |        1.00 |
| RTX 3090   | llama 8B Q6_K     |                 256 | Yes                | pp512    |        1927.92 |                 1970.42 |        1.02 |
| RTX 3090   | llama 8B Q6_K     |                 512 | Yes                | pp512    |        2088.83 |                 2116.00 |        1.01 |
| RTX 3090   | llama 8B Q8_0     |                  16 | Yes                | pp512    |         603.33 |                  645.74 |        1.07 |
| RTX 3090   | llama 8B Q8_0     |                  32 | Yes                | pp512    |         982.85 |                  984.64 |        1.00 |
| RTX 3090   | llama 8B Q8_0     |                  64 | Yes                | pp512    |        1430.97 |                 1456.70 |        1.02 |
| RTX 3090   | llama 8B Q8_0     |                 128 | Yes                | pp512    |        1856.46 |                 1876.76 |        1.01 |
| RTX 3090   | llama 8B Q8_0     |                 256 | Yes                | pp512    |        2325.70 |                 2321.93 |        1.00 |
| RTX 3090   | llama 8B Q8_0     |                 512 | Yes                | pp512    |        2471.43 |                 2497.23 |        1.01 |
| RX 6800    | llama 8B Q2_K_M   |                 512 | No                 | pp512    |         157.08 |                  163.49 |        1.04 |
| RX 6800    | llama 8B Q2_K_M   |                 512 | No                 | pp512    |         193.74 |                  194.98 |        1.01 |
| RX 6800    | llama 8B Q2_K_M   |                 512 | No                 | pp512    |         249.34 |                  243.46 |        0.98 |
| RX 6800    | llama 8B Q2_K_M   |                 512 | No                 | pp512    |         288.20 |                  277.75 |        0.96 |
…